update USAGE.md with central bo_create call
This is how I would envision the bo alloc call to work.

Having the liballoc core do the work to ask around which device
is able to allocate with a given set of caps/constraints makes
the application API easier and would allow to plug in things
like ION by just providing a liballoc driver backend on top of
them.