Add assertion hints query
There should be an easy way for applications to query
valid values for assertions.

This change adds a new assertion_hint_t structure and
the corresponding function for applications to query
the list of valid assertion values given a usage.

Among others, this function can be used to retrieve
the minimum and maximum surface size or enumerate the
supported formats.

Signed-off-by: Miguel A Vico Moya <mvicomoya@nvidia.com>
Reviewed-by: James Jones <jajones@nvidia.com>