Add allocation export support
This change adds a new function to export a
previously created allocation from a device to a
file descriptor and a raw byte stream of metadata.
There is no corresponding import function, because
exporting an allocation is the only useful
operation the allocater library can perform on an
allocation.  The intent is that exported
allocations be subsequently imported into
other libraries providing APIs which can perform
other useful operations on them.

Signed-off-by: James Jones <jajones@nvidia.com>
Reviewed-by: Miguel A Vico Moya <mvicomoya@nvidia.com>