Export a utility function to free capability sets
Various functions in the library allocate and
return a capability set or an array of capability
sets.  Freeing this is complicated because they
are not flat structures and contain const members
that are mildly tricky to pass into free() without
generating const correctness warnings.

To make things easier on programs, export a new
function: free_capability_sets() that will free an
array of capability sets of the specified size.
Also make use of the new utility function in the
test code and failure paths in the main library
where possible.

Signed-off-by: James Jones <jajones@nvidia.com>
Reviewed-by: Miguel A Vico Moya <mvicomoya@nvidia.com>